sched: stop rounding the kernel stack pointer through uint32_t
Three copies of

    uint32_t *kstack = (uint32_t *)((uint32_t)kstack_base + 0x4000);

truncate a pointer to 32 bits and cast it back.  On i386 that is a no-op,
which is why it has never mattered; the tests/ host build compiles these
sources natively, where it produces a garbage pointer and the first write
to the new thread's stack segfaults.

uintptr_t is the same width on i386 and correct everywhere else.

Kernel rebuilt and links; the host testsuite now gets past thread
creation and runs all 151 cases.

+++ sys/arch/i386/sched.c
@@ -132,7 +132,11 @@ int sched_fork_thread(process_t *proc, void *parent_regs) {
     }
 
     // Stack is at top of these pages (16KB = 0x4000)
-    uint32_t *kstack = (uint32_t *)((uint32_t)kstack_base + 0x4000);
+    /* uintptr_t, not uint32_t: rounding a pointer through a 32-bit integer
+     * truncates it on any host where pointers are wider, which is exactly
+     * what the tests/ host build is.  Identical on i386, where uintptr_t
+     * is 32 bits. */
+    uint32_t *kstack = (uint32_t *)((uintptr_t)kstack_base + 0x4000);
     t->kstack_base = (uintptr_t)kstack_base;
     t->kstack_top = (uintptr_t)kstack;
     t->kstack_units = 4;
@@ -227,7 +231,11 @@ int sched_clone_thread(process_t *proc, void *parent_regs, uint32_t tls_base,
         return -1;
     }
 
-    uint32_t *kstack = (uint32_t *)((uint32_t)kstack_base + 0x4000);
+    /* uintptr_t, not uint32_t: rounding a pointer through a 32-bit integer
+     * truncates it on any host where pointers are wider, which is exactly
+     * what the tests/ host build is.  Identical on i386, where uintptr_t
+     * is 32 bits. */
+    uint32_t *kstack = (uint32_t *)((uintptr_t)kstack_base + 0x4000);
     t->kstack_base = (uintptr_t)kstack_base;
     t->kstack_top = (uintptr_t)kstack;
     t->kstack_units = 4;
@@ -336,7 +344,11 @@ thread_t *sched_create_thread(process_t *proc, void (*entry_point)(void*), void 
         t->state = THREAD_ZOMBIE;
         return NULL;
     }
-    uint32_t *kstack = (uint32_t *)((uint32_t)kstack_base + 0x4000);
+    /* uintptr_t, not uint32_t: rounding a pointer through a 32-bit integer
+     * truncates it on any host where pointers are wider, which is exactly
+     * what the tests/ host build is.  Identical on i386, where uintptr_t
+     * is 32 bits. */
+    uint32_t *kstack = (uint32_t *)((uintptr_t)kstack_base + 0x4000);
     t->kstack_base  = (uintptr_t)kstack_base;
     t->kstack_top   = (uintptr_t)kstack;
     t->kstack_units = 4;
